+++ EZRecipes/app/src/main/java/com/abhiek/ezrecipes/ui/util/Accordion.kt
@@ -1,4 +1,4 @@
-package com.abhiek.ezrecipes.ui.profile
+package com.abhiek.ezrecipes.ui.util
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.expandVertically
