+++ EZRecipes/app/src/main/java/com/abhiek/ezrecipes/ui/util/CheckboxRow.kt
@@ -1,4 +1,4 @@
-package com.abhiek.ezrecipes.ui.search
+package com.abhiek.ezrecipes.ui.util
 
 import androidx.annotation.StringRes
 import androidx.compose.foundation.clickable
